Refactor Tap event parsing into common util code
